+++ EventSubscriber/ModuleEventSubscriber.php
@@ -8,11 +8,18 @@
  * file that was distributed with this source code.
  */
 
-
 namespace Austral\AdminBundle\EventSubscriber;
 
+use Austral\AdminBundle\Configuration\AdminConfiguration;
 use Austral\AdminBundle\Event\ModuleEvent;
+use Austral\EntityBundle\Entity\Interfaces\FilterByDomainInterface;
+use Austral\EntityFileBundle\File\Link\Generator;
+use Austral\HttpBundle\Entity\Interfaces\DomainInterface;
+use Austral\HttpBundle\Services\DomainsManagement;
+use Austral\ToolsBundle\AustralTools;
+use Doctrine\ORM\Query\QueryException;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * Austral Http EventSubscriber.
@@ -23,9 +30,24 @@ class ModuleEventSubscriber implements EventSubscriberInterface
 {
 
   /**
-   * @var bool
+   * @var TranslatorInterface
    */
-  protected bool $editMyAccount = false;
+  protected TranslatorInterface $translator;
+
+  /**
+   * @var DomainsManagement
+   */
+  protected DomainsManagement $domains;
+
+  /**
+   * @var AdminConfiguration
+   */
+  protected AdminConfiguration $adminConfiguration;
+
+  /**
+   * @var Generator
+   */
+  protected Generator $fileLinkGenerator;
 
   /**
    * @return array[]
@@ -33,8 +55,27 @@ class ModuleEventSubscriber implements EventSubscriberInterface
   public static function getSubscribedEvents(): array
   {
     return [
-      ModuleEvent::EVENT_AUSTRAL_MODULE_ADD     =>  ["moduleAdd", 1024],
+      ModuleEvent::EVENT_AUSTRAL_MODULE_ADD     =>  ["moduleAdd", 0],
     ];
+  }
+
+  /**
+   * @param TranslatorInterface $translator
+   * @param DomainsManagement $domains
+   * @param Generator $fileLinkGenerator
+   * @param AdminConfiguration $adminConfiguration
+   *
+   * @throws QueryException
+   */
+  public function __construct(TranslatorInterface $translator,
+    DomainsManagement $domains,
+    Generator $fileLinkGenerator,
+    AdminConfiguration $adminConfiguration)
+  {
+    $this->translator = $translator;
+    $this->domains = $domains->initialize();
+    $this->fileLinkGenerator = $fileLinkGenerator;
+    $this->adminConfiguration = $adminConfiguration;
   }
 
   /**
@@ -44,7 +85,53 @@ class ModuleEventSubscriber implements EventSubscriberInterface
    */
   public function moduleAdd(ModuleEvent $moduleEvent)
   {
+    if($moduleEvent->getModule()->isEntityModule())
+    {
+      $entityManager = $moduleEvent->getModule()->getEntityManager();
+      if(AustralTools::usedImplements($entityManager->getClass(), FilterByDomainInterface::class))
+      {
+        if($this->domains->getEnabledDomainWithoutVirtual() > 1) {
+          $moduleChange = false;
+          $domains = $this->domains->getDomains();
+          /** @var DomainInterface $domain */
+          foreach($domains as $domain)
+          {
+            if(!$domain->getIsVirtual())
+            {
+              $moduleChange = true;
+              $moduleEvent->getModules()->generateModuleByDomain(
+                $moduleEvent->getModule()->getModuleKey(),
+                $moduleEvent->getModule()->getModuleParameters(),
+                $domain,
+                $moduleEvent->getModule()
+              );
+            }
+          }
+          if($moduleChange)
+          {
+            $moduleEvent->getModules()->generateModuleByDomain(
+              $moduleEvent->getModule()->getModuleKey(),
+              $moduleEvent->getModule()->getModuleParameters(),
+              null,
+              $moduleEvent->getModule()
+            );
+            $moduleEvent->getModule()->setActionName("listChildrenModules");
+            $moduleEvent->getModule()->setPathActions(array());
+          }
+        }
+      }
+    }
+  }
 
+  /**
+   * @param $key
+   * @param array $parameters
+   *
+   * @return string
+   */
+  public function trans($key, array $parameters = array()): string
+  {
+    return $this->translator->trans($key, $parameters, "austral");
   }