+++ Changelog.md
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+### Version 3.1.0 (2022-09-17)
+* Moved all the Entities interfaces in bundle  : entity-bundle
+* Added manage with multi domain
+* Rewrited the Module generator
+* Generated admin module auto by domain if entity has filter by domain
+* Added new link selection for components and wysiwygs
+* Added new icons
+* Fix bug in the twig template
+
 ### Version 3.0.0 (2021-02-10)
 * Created new repository and new branch in to Github
 * Changed the project License to the MIT